refactor: 분리되어 있던 Folder 도메인 합치기
