+++ src/main/java/com/flytrap/rssreader/api/alert/domain/Alert.java
@@ -1,5 +1,7 @@
 package com.flytrap.rssreader.api.alert.domain;
 
+import com.flytrap.rssreader.api.account.domain.AccountId;
+import com.flytrap.rssreader.api.folder.domain.FolderId;
 import com.flytrap.rssreader.global.model.DefaultDomain;
 import com.flytrap.rssreader.global.model.Domain;
 import lombok.Builder;
@@ -12,13 +14,13 @@ import lombok.ToString;
 public class Alert  implements DefaultDomain {
 
     private final AlertId id;
-    private final Long memberId;
-    private final Long folderId;
+    private final AccountId memberId;
+    private final FolderId folderId;
     private final AlertPlatform alertPlatform;
     private final String webhookUrl;
 
     @Builder
-    protected Alert(AlertId id, Long memberId, Long folderId, AlertPlatform alertPlatform, String webhookUrl) {
+    protected Alert(AlertId id, AccountId memberId, FolderId folderId, AlertPlatform alertPlatform, String webhookUrl) {
         this.id = id;
         this.memberId = memberId;
         this.folderId = folderId;
