+++ src/main/java/com/flytrap/rssreader/api/folder/domain/AccessibleFolders.java
@@ -3,6 +3,6 @@ package com.flytrap.rssreader.api.folder.domain;
 import java.util.List;
 
 public record AccessibleFolders(
-    List<FolderDomain> privateFolders,
-    List<FolderDomain> sharedFolder
+    List<Folder> privateFolders,
+    List<Folder> sharedFolder
 ) { }
